requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Google APIs (if needed for cloud deliverables)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data processing
pandas>=2.1.0

# Web scraping (if needed)
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Add other dependencies as needed
